DataSender.cs and a bug fixed in TransportLayerCommunicator
git-svn-id: https://tashjik.googlecode.com/svn/trunk@84 830311b4-703b-0410-b43d-057cfd0b486a

+++ Tier2/ChordDataStore.cs
@@ -71,7 +71,7 @@ namespace Tashjik.Tier2
 		private Dictionary<byte[], Stream> dataHolder =
 			new Dictionary<byte[], Stream>();
 
-		public void putData(byte[] hashedKey, Stream data, UInt32 dataLength)
+		public void putData(byte[] hashedKey, Stream data, UInt64 dataLength)
 		{
 			dataHolder.Add(hashedKey, data);
 		}
